(qol)check envs in deploy script

+++ scripts/deploy.sh
@@ -1,9 +1,21 @@
 #!/bin/bash -e
+
+DEPLOY_SERVER="$TINCHO_DEPLOY_SERVER"
+DEPLOY_USER="$TINCHO_DEPLOY_USER"
+DEPLOY_PATH="$TINCHO_DEPLOY_PATH"
+
+if [[ -z "$DEPLOY_SERVER" || -z "$DEPLOY_USER" || -z "$DEPLOY_PATH" ]]; then
+  echo "Missing environment variables"
+  exit 1
+fi
+
 mkdir -p build
 go build -o build/server cmd/server/main.go
-scp build/server "$TINCHO_DEPLOY_USER@$TINCHO_DEPLOY_SERVER:$TINCHO_DEPLOY_PATH/server_new"
-ssh -T "$TINCHO_DEPLOY_USER@$TINCHO_DEPLOY_SERVER" << EOF
+
+scp build/server "$DEPLOY_USER@$DEPLOY_SERVER:$DEPLOY_PATH/server_new"
+# shellcheck disable=SC2087
+ssh -T "$DEPLOY_USER@$DEPLOY_SERVER" << EOF  
 sudo systemctl stop tincho
-mv "$TINCHO_DEPLOY_PATH/server_new" "$TINCHO_DEPLOY_PATH/server"
+mv "$DEPLOY_PATH/server_new" "$DEPLOY_PATH/server"
 sudo systemctl start tincho
 EOF
